requirements.txt
# Core
numpy==1.26.4
pandas>=2.0
regex>=2023.12.25

# Scientific stack pins for compatibility
scipy==1.10.1
scikit-learn==1.3.2

# Embeddings
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4

# LLM API
openai>=1.37.0
python-dotenv>=1.0.1

# Utils
tqdm>=4.66.4
requests>=2.32.3
